+++ bencher/__init__.py
@@ -18,6 +18,7 @@ from bencher.results.holoview_results.line_result import LineResult
 from bencher.results.holoview_results.scatter_result import ScatterResult
 from bencher.results.holoview_results.surface_result import SurfaceResult
 from bencher.results.holoview_results.table_result import TableResult
+from bencher.results.holoview_results.tabular_spec import TabularSpec
 from bencher.results.holoview_results.tabulator_result import TabulatorResult
 from bencher.results.holoview_results.xy_scatter_result import XYScatterResult, xy_scatter
 from bencher.results.volume_result import VolumeResult